Further work toward UMM
